requirements.txt
python-dotenv
chainlit
requests
uv